Minor cross-mod compatibility improvement
If another mod adds a non-recipe effect, such as increasing evolution or movement speed, to one of the technologies enabled by the RobotWorld mod, a null reference error occurs (because we try to enable a recipe of an effect that has no recipe).  The little existence check I added in a few places will ignore non-recipe effects, thus preventing the error.

I also added some tabs to clean up formatting for the recursive-blueprints if-then statement.